exodus/backend/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.2
django-cors-headers==3.11.0
django-phonenumber-field==6.1.0
django-phonenumbers==1.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
phonenumbers==8.12.43
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
